third_party/blink/web_tests/external/wpt/editing/whitespaces/delete-word-around-span_white-space=normal-expected.txt

This is a testharness.js-based test.
[FAIL] Ctrl/Cmd - Backspace when "abc[] <span>def</span> ghi" should get same innerText as when "abc[] def ghi"
  assert_equals: expected " def ghi" but got "def ghi"
Harness: the test ran to completion.
